pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.backends.legacy:build"

[project]
name = "election-chat"
version = "0.1.0"
description = "Chat with Côte d'Ivoire 2025 election results"
requires-python = ">=3.10"

dependencies = [
    # Extraction PDF
    "pdfplumber>=0.11",
    "pdf2image>=1.17",
    "pytesseract>=0.3.10",
    "Pillow>=10.0",
    "pymupdf>=1.24",
    # Données
    "pandas>=2.0",
    "duckdb>=0.10",
    # LLM
    "anthropic>=0.30",
    # App
    "streamlit>=1.35",
    "plotly>=5.20",
    # RAG (Level 2)
    "sentence-transformers>=3.0",
    "faiss-cpu>=1.8",
    # Utils
    "python-dotenv>=1.0",
    "requests>=2.31",
    "tqdm>=4.66",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0",
    "pytest-cov>=5.0",
    "ruff>=0.4",
]

[project.scripts]
ingest = "ingestion.pipeline:main"

[tool.ruff]
line-length = 100
target-version = "py310"

[tool.pytest.ini_options]
testpaths = ["tests", "evals"]
python_files = ["test_*.py"]
